+++ languagetool-language-modules/fr/src/test/java/org/languagetool/rules/fr/QuestionWhitespaceStrictRuleTest.java
@@ -18,10 +18,6 @@
  */
 package org.languagetool.rules.fr;
 
-import static org.junit.Assert.assertEquals;
-
-import java.io.IOException;
-
 import org.junit.Test;
 import org.languagetool.JLanguageTool;
 import org.languagetool.Language;
@@ -29,15 +25,19 @@ import org.languagetool.TestTools;
 import org.languagetool.language.French;
 import org.languagetool.rules.RuleMatch;
 
+import java.io.IOException;
+
+import static org.junit.Assert.assertEquals;
+
 /**
  * @author Marcin Miłkowski
  */
-public class QuestionWhitespaceRuleTest {
+public class QuestionWhitespaceStrictRuleTest {
 
     @Test
     public final void testRule() throws IOException {
-    	Language french = new French();
-      QuestionWhitespaceRule rule = new QuestionWhitespaceRule(TestTools.getEnglishMessages(), french);
+      Language french = new French();
+      QuestionWhitespaceStrictRule rule = new QuestionWhitespaceStrictRule(TestTools.getEnglishMessages(), french);
       RuleMatch[] matches;
       JLanguageTool langTool = new JLanguageTool(french);
       
@@ -50,18 +50,28 @@ public class QuestionWhitespaceRuleTest {
       // errors:
       matches = rule.match(langTool.getAnalyzedSentence("C'est vrai!"));
       assertEquals(1, matches.length);
+      matches = rule.match(langTool.getAnalyzedSentence("C'est vrai !"));
+      assertEquals(1, matches.length);
+      matches = rule.match(langTool.getAnalyzedSentence("Qu'est ce que c'est ?"));
+      assertEquals(1, matches.length);
       matches = rule.match(langTool.getAnalyzedSentence("Qu'est ce que c'est?"));
       assertEquals(1, matches.length);
       matches = rule.match(langTool.getAnalyzedSentence("L'enjeu de ce livre est donc triple: philosophique;"));
       assertEquals(2, matches.length);
+      assertEquals(1, rule.match(langTool.getAnalyzedSentence("Bonjour : )")).length);
       matches = rule.match(langTool.getAnalyzedSentence("L'enjeu de ce livre est donc triple: philosophique ;"));
-      assertEquals(1, matches.length);
+      assertEquals(2, matches.length);
       // check match positions:
+      assertEquals(2, matches.length);
       assertEquals(29, matches[0].getFromPos());
       assertEquals(36, matches[0].getToPos());
+      assertEquals(50, matches[1].getFromPos());
+      assertEquals(52, matches[1].getToPos());
       //guillemets
+      matches = rule.match(langTool.getAnalyzedSentence("Le guillemet ouvrant est suivi d'un espace insécable : « mais le lieu [...] et le guillemet fermant est précédé d'un espace insécable : [...] littérature »."));
+      assertEquals(2, matches.length);
       matches = rule.match(langTool.getAnalyzedSentence("Le guillemet ouvrant est suivi d'un espace insécable : «mais le lieu [...] et le guillemet fermant est précédé d'un espace insécable : [...] littérature»."));
-      assertEquals(1, matches.length);
+      assertEquals(2, matches.length);
     }
     
 }